% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasta.R
\name{call_pasta_addin}
\alias{call_pasta_addin}
\title{Call the addin--just a wrapper for \code{pasta()} with a timestamped filename}
\usage{
call_pasta_addin()
}
\description{
Call the addin--just a wrapper for \code{pasta()} with a timestamped filename
}
