% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasta.R
\name{pasta}
\alias{pasta}
\title{Create a new file to dump text into}
\usage{
pasta(filename, dir = Sys.getenv("PASTA_DIR"), edit = TRUE)
}
\arguments{
\item{filename}{name of the file to save, including extension}

\item{dir}{Directory to save to, defaults to PASTA_DIR value}

\item{edit}{passed to \code{rmarkdown::draft()}}
}
\description{
Create a new file to dump text into
}
