% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work_on.R
\name{work_on}
\alias{work_on}
\title{Switch projects, possibly in a new session}
\usage{
work_on(projectname, new = FALSE, project_home = Sys.getenv("PROJECT_HOME"))
}
\arguments{
\item{projectname}{Name of project withing project_home}

\item{new}{if TRUE, open the project in a new session (and window) of Rstudio}

\item{project_home}{projectname must be within this directory,
defaults to PROJECT_HOME environment variable.}
}
\description{
Switch projects, possibly in a new session
}
